Word Counter Function Push
Added several functions to count the number of times that words appear in a given essay.

'compileWordsRunner' is a function that takes in an essay and formats it into an array so that it may be fed into the 'compileWords' function.

'compileWords' receives an array of strings, and an empty dictionary to begin, then will call itself recursively until it has looped through every element of the array of strings. It uses the 'endOfArray' function to determine if it has reached the end. It returns the dictionary when it is complete.

'endOfArray' is a helper to determine if the 'compileWords' function has reached the end of an array, and will end the recursion if it has.