Added var 'dict'
Created a variable called 'dict' to store the dictionary of words in the essay, and the value of times they appear. This is an improvement from the previous structure of the code where these words and values were not being stored anywhere, but just being returned by the 'compileWordsRunner' function. 

Some comments and structure of the functions were changed to reflect the change mentioned above.