Added MaxRepeat and Synonym Getters
*ADDED VARIABLES*
'maxRepeat' is a variable that was added to set a threshold for the maximum number of times a user wants a word to be repeated before it is flagged for replacement. This feature will be implemented in later functions.

'commonWords' is an array of common words that will be ignored when when words are being flagged for synonyms. The list will be further fleshed out at a later point when the project is in a more finalized state. 

*ADDED FUNCTIONS* 
'setMaxRepeat' is a function that will take a user input and set the maxRepeat variable equal to it. This will be fully implemented later with the addition of a GUI.

'getNouns', 'getVerbs', and 'getAdjectives' are all functions that return an array of synonyms of their respective part of speech for a given word. They will later be implemented in a master function that lists replacement options for a given word.