Frontend Added for all Non-API Functions
*NEW DIVS*
'output', 'container', and 'highlights' are divs that will have functionality added later to highlight the words that are repeated too many times.

*CHANGED DIVS*
'Input' replaced with 'Container'
A button was added to perform the 'setEssayVal' function.
An output textarea has been added. 
An input area for the max repeats has been added.

*NEW VARS*
'RepeatedWords' is used to store the values of 'flaggedWords' as a string. 

*NEW FUNCTIONS*
'setEssayVal' takes all of the user input present on the page, pushes it through the existing functions, and outputs the words that are repeated more than the maximum number of times, in the output div. 
'flaggedWordsToString' gets the values of the elements of flaggedWords and saves them to 'flaggedWords' in the format "word, word, word, etc...".

*CHANGED FUNCTIONS*
'setMaxRepeat' now takes the input value from the user, and saves it to maxRepeat.
'wordRepeatChecker' now excludes strings that are included in 'commonWords'
'compileWordsRunner' now replaces all newlines with spaces. This prevents an error that was occurring that caused 'dict' to save words adjacent to new lines separately from identical words that are not found next to new lines.