Add files via upload
*VARIABLES ADDED* 
The 'synonymDict' variale is used to store the objects returned by '$getWords' so they may be later used by the 'getSynonyms' function.

The 'finalOutput' variable is the final string that is created as output to tell the user the words that have been repeated, how many times they have been repeated, and their synonyms.

*FUNCTIONS ADDED*
The 'getSynonyms' function assigns the values of synonyms from the synonymDict values, organizes them in the 'finalOutput' string. The format shows the word, number of time it is repeated, and the respective synonyms for nouns, verbs, and adjectives. The 'EssayOutput' value is then printed to the EssayOutput box.

*FUNCTIONS CHANGED*
'getNouns', 'getVerbs', and 'getAdjectives' all have been changed to prevent undefined words from causing errors when the program runs.

*Cosmetic Changes*
The text boxes and buttons have been centered, and the background color has been changed.